requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# CLI and API
click>=8.1.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database and memory
sqlalchemy>=2.0.0
alembic>=1.12.0
chromadb>=0.4.0

# AI backends - Cloud
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
httpx>=0.24.0
aiofiles>=23.0.0
python-multipart>=0.0.6
pyyaml>=6.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# NOTE: llama-cpp-python removed from requirements due to installation complexity
# See requirements-local.txt for local LLM dependencies
# Or use cloud mode with API keys (recommended for quick start)
